Ultimo cambio del backend por hoy

+++ Backend/src/main/java/com/apiweb/backend/Model/UsuariosModel.java
@@ -21,7 +21,7 @@ public class UsuariosModel {
     private String nombre;
     private String correo;
     private String contrasena;
-    private Integer telefono;
+    private Long telefono;
     private TipoUsuario tipo;
     private String pregunta_seguridad;
     private ArrayList<Calificaciones> calificaciones = new ArrayList<Calificaciones>();
